Fix broken skill mode switching offense->defense
Problem was: when a text fragment goes active it would be added to the battlemanager's set of active fragments
this included the fragment which was associated with a skill.
However, the battle manager was only intended to hold on to text fragments for defense.
Also refactored mode switching to be handled by a state machine.