Store fragment timers when text fragments are created by NPC
Because enemy npc attack consists of playing the animation
then waiting for some delay before creating the actual fragment,
if the npc is destroyed after the animation but before the
fragment gets sent, then the timer will fire after the npc
is removed.  Now we save those timers in an array so they
can be cleared when the NPC is destroyed.