Refactored behavior for projectiles hitting
The behavior of a collision happening which triggers another set of
effects was previously stuffed into BattleNPCProjectile.
I believe that component and it's system should only handle movement of
the projectiles.

On second thought, perhaps it's ok for that component to handle both
jobs but there should still be a separate system which handles
triggering of teh effect set.  For now, this refactor is good