Update status bar on damage
Change damage to use an effect system
effect system has a queue of effects
effects are functions which take an entity as an arg
effects are pushed into the queue rather than directly affecting (ha)
the entity.  When the effect system runs, it pops things off the queue
and calls them with the entity.
When this happens, we can set the status to dirty
then the entity status system runs and checks if the status is dirty
if so, then it'll rerender the status view.

It seems bigger and more clunky, but ultimately reduces dependencies bigly.
Otherwise, we'd have the status system listening for an "UpdateStatus"
event, and all kinds of random components and effects and who knows what
else would be triggering that event.  The event may be triggered
multiple times within a frame even, causing multiple unnecessary
rerenders.

With this setup, only the effect system can (should) actually make
updates to entities (those which have the BattleEffectable comopnent
anyway).  The trend seems to be funneling side effects down to a single
point of execution.  This makes it easier to keep track of when and what
things are happening to any given entity.  Makes it so things are only
happening on frame (system evaluation).