Fix issue with views lingering on after scene
So the problem was: I was declaring _childViews as an array as an
attribute on the backbone view.  Now if I remember, I had the same
problem with crafty components.  When declaring a component with some
attribute that is a pointer (array, obj, etc) then all instances of that
component end up sharing the same pointer and ultimately the same
object.  Clearly this is not what I want.

What I need to do is to declare the _childViews attribute in the
initializer as
this._childViews = [];
which would ensure that differnt instances of the view get their own set
of child views.