Tutorial stages more controlled by tutorial system
Tutorial system is what ends the scene rather than binding in some
random system.  Previously, you beat the monster to complete the simple
tutorial.  Now the tutorial ends and that's what completes the tutorial,
as well as all the others.

Add a tutorial step which waits for an event, this triggers moving on to
the next tutorial step.